+++ src/main/java/org/folio/entlinks/controller/AuthoritySourceFilesController.java
@@ -8,6 +8,7 @@ import org.folio.entlinks.controller.delegate.AuthoritySourceFileServiceDelegate
 import org.folio.entlinks.domain.dto.AuthoritySourceFileDto;
 import org.folio.entlinks.domain.dto.AuthoritySourceFileDtoCollection;
 import org.folio.entlinks.domain.dto.AuthoritySourceFilePatchDto;
+import org.folio.entlinks.domain.dto.AuthoritySourceFilePostDto;
 import org.folio.entlinks.rest.resource.AuthoritySourceFileApi;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,8 +23,8 @@ public class AuthoritySourceFilesController implements AuthoritySourceFileApi {
   private final AuthoritySourceFileServiceDelegate delegate;
 
   @Override
-  public ResponseEntity<AuthoritySourceFileDto> createAuthoritySourceFile(AuthoritySourceFileDto authoritySourceFile) {
-    var created = delegate.createAuthoritySourceFile(authoritySourceFile);
+  public ResponseEntity<AuthoritySourceFileDto> createAuthoritySourceFile(AuthoritySourceFilePostDto postDto) {
+    var created = delegate.createAuthoritySourceFile(postDto);
     return ResponseEntity.status(HttpStatus.CREATED).body(created);
   }