+++ src/main/java/org/folio/entlinks/domain/entity/AuthoritySourceFile.java
@@ -62,6 +62,15 @@ public class AuthoritySourceFile extends MetadataEntity implements Persistable<U
   @OneToMany(mappedBy = "authoritySourceFile", fetch = FetchType.LAZY)
   private Set<Authority> authorities = new LinkedHashSet<>();
 
+  @Column(name = "sequence_name", length = 50)
+  private String sequenceName;
+
+  @Column(name = "selectable")
+  private boolean selectable = false;
+
+  @Column(name = "hrid_start_number")
+  private Integer hridStartNumber;
+
   @Transient
   private boolean isNew = true;
 
@@ -80,6 +89,9 @@ public class AuthoritySourceFile extends MetadataEntity implements Persistable<U
     this.authorities = Optional.ofNullable(other.authorities).orElse(Set.of()).stream()
         .map(Authority::new)
         .collect(Collectors.toSet());
+    this.sequenceName = other.sequenceName;
+    this.selectable = other.selectable;
+    this.hridStartNumber = other.hridStartNumber;
   }
 
   public void addCode(AuthoritySourceFileCode code) {