+++ src/main/java/org/folio/entlinks/service/authority/AuthoritySourceFileService.java
@@ -6,27 +6,38 @@ import static org.folio.entlinks.utils.ServiceUtils.initId;
 import java.util.List;
 import java.util.Objects;
 import java.util.UUID;
+import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
 import lombok.AllArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.lang3.StringUtils;
 import org.folio.entlinks.controller.converter.AuthoritySourceFileMapper;
 import org.folio.entlinks.domain.entity.AuthoritySourceFile;
+import org.folio.entlinks.domain.entity.AuthoritySourceFileCode;
+import org.folio.entlinks.domain.entity.AuthoritySourceType;
 import org.folio.entlinks.domain.repository.AuthoritySourceFileRepository;
 import org.folio.entlinks.exception.AuthoritySourceFileNotFoundException;
 import org.folio.entlinks.exception.RequestBodyValidationException;
+import org.folio.spring.FolioExecutionContext;
+import org.folio.spring.FolioModuleMetadata;
 import org.folio.spring.data.OffsetRequest;
 import org.folio.tenant.domain.dto.Parameter;
 import org.springframework.data.domain.Page;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @AllArgsConstructor
 @Log4j2
 public class AuthoritySourceFileService {
 
+  private static final String AUTHORITY_SEQUENCE_NAME_TEMPLATE = "hrid_authority_local_file_%s_seq";
   private final AuthoritySourceFileRepository repository;
   private final AuthoritySourceFileMapper mapper;
+  private final JdbcTemplate jdbcTemplate;
+  private final FolioModuleMetadata moduleMetadata;
+  private final FolioExecutionContext folioExecutionContext;
 
   public Page<AuthoritySourceFile> getAll(Integer offset, Integer limit, String cql) {
     log.debug("getAll:: Attempts to find all AuthoritySourceFile by [offset: {}, limit: {}, cql: {}]", offset, limit,
@@ -55,14 +66,13 @@ public class AuthoritySourceFileService {
     return repository.findByName(name).orElse(null);
   }
 
+  @Transactional
   public AuthoritySourceFile create(AuthoritySourceFile entity) {
     log.debug("create:: Attempting to create AuthoritySourceFile [entity: {}]", entity);
 
-    initId(entity);
+    validateOnCreate(entity);
 
-    for (var code : entity.getAuthoritySourceFileCodes()) {
-      code.setAuthoritySourceFile(entity);
-    }
+    initOnCreate(entity);
 
     return repository.save(entity);
   }
@@ -91,6 +101,46 @@ public class AuthoritySourceFileService {
     } else {
       throw new RequestBodyValidationException("Cannot delete Authority source file with source 'folio'",
           List.of(new Parameter("source").value(authoritySourceFile.getSource())));
+    }
+  }
+
+  public long getNextSequenceNumber(String sequenceName) {
+    return repository.getNextSequenceNumber(sequenceName);
+  }
+
+  public void createSequence(String sequenceName, int startNumber) {
+    var command = String.format("""
+        CREATE SEQUENCE %s MINVALUE %d INCREMENT BY 1 OWNED BY %s.authority_source_file.sequence_name
+        """,
+        sequenceName, startNumber, moduleMetadata.getDBSchemaName(folioExecutionContext.getTenantId()));
+    jdbcTemplate.execute(command);
+  }
+
+  private void validateOnCreate(AuthoritySourceFile entity) {
+    if (!AuthoritySourceType.LOCAL.getValue().equals(entity.getSource())) {
+      throw new RequestBodyValidationException("Only Authority Source File with source Local can be created",
+          List.of(new Parameter("source").value(entity.getSource())));
+    }
+
+    if (entity.getAuthoritySourceFileCodes().size() != 1) {
+      var codes = entity.getAuthoritySourceFileCodes().stream()
+          .map(AuthoritySourceFileCode::getCode)
+          .collect(Collectors.joining(","));
+      throw new RequestBodyValidationException("Authority Source File with source Local should have only one prefix",
+          List.of(new Parameter("code").value(codes)));
+    }
+  }
+
+  private void initOnCreate(AuthoritySourceFile entity) {
+    initId(entity);
+
+    var sourceFileCode = entity.getAuthoritySourceFileCodes().iterator().next();
+    sourceFileCode.setAuthoritySourceFile(entity);
+    var sequenceName = String.format(AUTHORITY_SEQUENCE_NAME_TEMPLATE, sourceFileCode.getCode());
+    entity.setSequenceName(sequenceName);
+
+    if (entity.getHridStartNumber() == null) {
+      entity.setHridStartNumber(1);
     }
   }