+++ src/test/java/org/folio/entlinks/controller/AuthoritySourceFilesControllerIT.java
@@ -24,6 +24,8 @@ import java.util.UUID;
 import org.folio.entlinks.domain.dto.AuthoritySourceFileDto;
 import org.folio.entlinks.domain.dto.AuthoritySourceFileDto.SourceEnum;
 import org.folio.entlinks.domain.dto.AuthoritySourceFileDtoCollection;
+import org.folio.entlinks.domain.dto.AuthoritySourceFilePostDto;
+import org.folio.entlinks.domain.dto.AuthoritySourceFilePostDtoHridManagement;
 import org.folio.entlinks.domain.entity.AuthoritySourceFile;
 import org.folio.entlinks.domain.entity.AuthoritySourceFileCode;
 import org.folio.entlinks.exception.AuthoritySourceFileNotFoundException;
@@ -53,8 +55,7 @@ class AuthoritySourceFilesControllerIT extends IntegrationTestBase {
   private static final Integer[] SOURCE_FILE_CODE_IDS = new Integer[] {1, 2, 3};
   private static final String[] SOURCE_FILE_CODES = new String[] {"code1", "code2", "code3"};
   private static final String[] SOURCE_FILE_NAMES = new String[] {"name1", "name2", "name3"};
-  private static final SourceEnum[] SOURCE_FILE_SOURCES =
-    new SourceEnum[] {SourceEnum.LOCAL, SourceEnum.LOCAL, SourceEnum.FOLIO};
+  private static final String[] SOURCE_FILE_SOURCES = {"local", "local", "folio"};
   private static final String[] SOURCE_FILE_TYPES = new String[] {"type1", "type2", "type3"};
   private static final String[] SOURCE_FILE_URLS = new String[] {"baseUrl1", "baseUrl2", "baseUrl3"};
 
@@ -79,6 +80,7 @@ class AuthoritySourceFilesControllerIT extends IntegrationTestBase {
     tryGet(authoritySourceFilesEndpoint())
       .andExpect(status().isOk())
       .andExpect(jsonPath("totalRecords", is(createdEntities.size())))
+      .andExpect(jsonPath("authoritySourceFiles[1].hridManagement.startNumber", is(2)))
       .andExpect(jsonPath("authoritySourceFiles[0].metadata", notNullValue()))
       .andExpect(jsonPath("authoritySourceFiles[0].metadata.createdDate", notNullValue()))
       .andExpect(jsonPath("authoritySourceFiles[0].metadata.updatedDate", notNullValue()))
@@ -115,6 +117,7 @@ class AuthoritySourceFilesControllerIT extends IntegrationTestBase {
 
     doGet(authoritySourceFilesEndpoint(sourceFile.getId()))
         .andExpect(jsonPath("name", is(sourceFile.getName())))
+        .andExpect(jsonPath("hridManagement.startNumber", is(1)))
         .andExpect(jsonPath("metadata.createdDate", notNullValue()))
         .andExpect(jsonPath("metadata.createdByUserId", is(USER_ID)));
   }
@@ -146,20 +149,27 @@ class AuthoritySourceFilesControllerIT extends IntegrationTestBase {
   void createAuthoritySourceFile_positive_entityCreated() throws Exception {
     assumeTrue(databaseHelper.countRows(DatabaseHelper.AUTHORITY_SOURCE_FILE_TABLE, TENANT_ID) == 0);
 
-    var dto = new AuthoritySourceFileDto("name", List.of("code"), "type", SourceEnum.FOLIO)
-      .baseUrl("url");
+    var id = UUID.randomUUID();
+    var dto = new AuthoritySourceFilePostDto()
+        .id(id).name("name").code("code").type("type").baseUrl("url").selectable(true)
+        .hridManagement(new AuthoritySourceFilePostDtoHridManagement().startNumber(10));
+    var expectedSequenceName = "hrid_authority_local_file_code_seq";
 
     tryPost(authoritySourceFilesEndpoint(), dto)
       .andExpect(status().isCreated())
       .andExpect(jsonPath("name", is(dto.getName())))
-      .andExpect(jsonPath("source", is(dto.getSource().getValue())))
-      .andExpect(jsonPath("codes", is(dto.getCodes())))
+      .andExpect(jsonPath("source", is("local")))
+      .andExpect(jsonPath("codes", is(List.of(dto.getCode()))))
+      .andExpect(jsonPath("selectable", is(true)))
       .andExpect(jsonPath("metadata.createdDate", notNullValue()))
       .andExpect(jsonPath("metadata.updatedDate", notNullValue()))
       .andExpect(jsonPath("metadata.updatedByUserId", is(USER_ID)))
       .andExpect(jsonPath("metadata.createdByUserId", is(USER_ID)));
 
     assertEquals(1, databaseHelper.countRows(DatabaseHelper.AUTHORITY_SOURCE_FILE_TABLE, TENANT_ID));
+    assertEquals(expectedSequenceName, databaseHelper.queryAuthoritySourceFileSequenceName(TENANT_ID, id));
+    assertEquals(dto.getHridManagement().getStartNumber(),
+        databaseHelper.queryAuthoritySourceFileSequenceStartNumber(expectedSequenceName));
   }
 
   @Test
@@ -167,8 +177,8 @@ class AuthoritySourceFilesControllerIT extends IntegrationTestBase {
   void createAuthoritySourceFile_negative_existedName() throws Exception {
     var createdEntities = createAuthoritySourceTypes();
 
-    var dto = new AuthoritySourceFileDto(createdEntities.get(0).getName(),
-      List.of("code"), "type", SourceEnum.FOLIO).baseUrl("url");
+    var dto = new AuthoritySourceFilePostDto(createdEntities.get(0).getName(), "code")
+        .baseUrl("url").type("type");
 
     tryPost(authoritySourceFilesEndpoint(), dto)
       .andExpect(status().isUnprocessableEntity())
@@ -184,8 +194,8 @@ class AuthoritySourceFilesControllerIT extends IntegrationTestBase {
   void createAuthoritySourceFile_negative_existedUrl() throws Exception {
     var createdEntities = createAuthoritySourceTypes();
 
-    var dto = new AuthoritySourceFileDto("new name",
-      List.of("code"), "type", SourceEnum.FOLIO).baseUrl(createdEntities.get(0).getBaseUrl());
+    var dto = new AuthoritySourceFilePostDto("new name", "code")
+        .baseUrl(createdEntities.get(0).getBaseUrl()).type("type");
 
     tryPost(authoritySourceFilesEndpoint(), dto)
       .andExpect(status().isUnprocessableEntity())
@@ -201,8 +211,7 @@ class AuthoritySourceFilesControllerIT extends IntegrationTestBase {
   void createAuthoritySourceFile_negative_existedCode() throws Exception {
     var createdEntities = createAuthoritySourceTypes();
 
-    var dto = new AuthoritySourceFileDto("new name",
-      List.of("code1", "code5"), "type", SourceEnum.FOLIO).baseUrl("new url");
+    var dto = new AuthoritySourceFilePostDto("new name", "code1").baseUrl("new url").type("type");
 
     tryPost(authoritySourceFilesEndpoint(), dto)
       .andExpect(status().isUnprocessableEntity())
@@ -216,7 +225,7 @@ class AuthoritySourceFilesControllerIT extends IntegrationTestBase {
   @Test
   @DisplayName("POST: return 422 for authority source file without name")
   void createAuthoritySourceFile_negative_entityWithoutNameNotCreated() throws Exception {
-    var dto = new AuthoritySourceFileDto(null, List.of("code"), "type", SourceEnum.LOCAL);
+    var dto = new AuthoritySourceFilePostDto(null, "code").type("type");
 
     tryPost(authoritySourceFilesEndpoint(), dto)
         .andExpect(status().isUnprocessableEntity())
@@ -232,8 +241,7 @@ class AuthoritySourceFilesControllerIT extends IntegrationTestBase {
   @Test
   @DisplayName("PATCH: partially update Authority Source File")
   void updateAuthoritySourceFilePartially_positive_entityGetUpdated() throws Exception {
-    var dto = new AuthoritySourceFileDto("name", List.of("code1"), "type", SourceEnum.FOLIO)
-      .baseUrl("url");
+    var dto = new AuthoritySourceFilePostDto("name", "code1").type("type").baseUrl("url");
 
     doPost(authoritySourceFilesEndpoint(), dto);
     var existingAsString = doGet(authoritySourceFilesEndpoint()).andReturn().getResponse().getContentAsString();
@@ -325,9 +333,10 @@ class AuthoritySourceFilesControllerIT extends IntegrationTestBase {
     var entity = new AuthoritySourceFile();
     entity.setId(SOURCE_FILE_IDS[i]);
     entity.setName(SOURCE_FILE_NAMES[i]);
-    entity.setSource(SOURCE_FILE_SOURCES[i].getValue());
+    entity.setSource(SOURCE_FILE_SOURCES[i]);
     entity.setType(SOURCE_FILE_TYPES[i]);
     entity.setBaseUrl(SOURCE_FILE_URLS[i]  + "/");
+    entity.setHridStartNumber(i + 1);
 
     var code = prepareAuthoritySourceFileCode(i);
     entity.setCreatedDate(Timestamp.from(Instant.parse(CREATED_DATE)));