+++ src/test/java/org/folio/entlinks/service/authority/AuthoritySourceFileServiceTest.java
@@ -25,7 +25,6 @@ import org.folio.entlinks.exception.RequestBodyValidationException;
 import org.folio.spring.test.type.UnitTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -113,15 +112,47 @@ class AuthoritySourceFileServiceTest {
     var code = new AuthoritySourceFileCode();
     var entity = new AuthoritySourceFile();
     entity.setAuthoritySourceFileCodes(Set.of(code));
-    var expected = new AuthoritySourceFile();
-    when(repository.save(any(AuthoritySourceFile.class))).thenReturn(expected);
-    var argumentCaptor = ArgumentCaptor.forClass(AuthoritySourceFile.class);
+    entity.setSource("local");
+    var expected = new AuthoritySourceFile(entity);
+    when(repository.save(any(AuthoritySourceFile.class))).thenAnswer(i -> i.getArguments()[0]);
 
     var created = service.create(entity);
 
+    expected.setId(created.getId());
     assertThat(created).isEqualTo(expected);
-    verify(repository).save(argumentCaptor.capture());
-    assertThat(argumentCaptor.getValue().getId()).isNotNull();
+  }
+
+  @Test
+  void shouldNotBePossibleToCreateAuthoritySourceFileOfSourceFolio() {
+    var code = new AuthoritySourceFileCode();
+    var entity = new AuthoritySourceFile();
+    entity.setAuthoritySourceFileCodes(Set.of(code));
+    entity.setSource("folio");
+
+    var thrown = assertThrows(RequestBodyValidationException.class, () -> service.create(entity));
+
+    verifyNoInteractions(repository);
+    assertThat(thrown.getInvalidParameters()).hasSize(1);
+    assertThat(thrown.getInvalidParameters().get(0).getKey()).isEqualTo("source");
+    assertThat(thrown.getInvalidParameters().get(0).getValue()).isEqualTo(entity.getSource());
+  }
+
+  @Test
+  void shouldNotBePossibleToCreateLocalAuthoritySourceWithMoreThanOneCode() {
+    var code1 = new AuthoritySourceFileCode();
+    code1.setCode("code1");
+    var code2 = new AuthoritySourceFileCode();
+    code2.setCode("code2");
+    var entity = new AuthoritySourceFile();
+    entity.setAuthoritySourceFileCodes(Set.of(code1, code2));
+    entity.setSource("local");
+
+    var thrown = assertThrows(RequestBodyValidationException.class, () -> service.create(entity));
+
+    verifyNoInteractions(repository);
+    assertThat(thrown.getInvalidParameters()).hasSize(1);
+    assertThat(thrown.getInvalidParameters().get(0).getKey()).isEqualTo("code");
+    assertThat(thrown.getInvalidParameters().get(0).getValue()).contains(code1.getCode()).contains(code2.getCode());
   }
 
   @Test